ALF-3845 - Справочник "Настройки АРМ" заново импортируется при каждом старте системы
git-svn-id: http://svn.datateh.ru/alfrescoecm/Business%20Platform/trunk@11771 697f8a5b-bb4b-4f4c-ab90-ff1ad96ece97

+++ lecm-dictionary/api/src/main/java/ru/it/lecm/dictionary/beans/XMLImportBeanImpl.java
@@ -273,7 +273,13 @@ public class XMLImportBeanImpl implements XMLImportBean {
 				    }
 			    } else if (!associationDefinition.isTargetMany()) {
 				    for (AssociationRef existingAssoc : existingAssocs) {
-					    nodeService.removeAssociation(parent, existingAssoc.getTargetRef(), assocType);
+                        if (!targetRef.equals(existingAssoc.getTargetRef())) {
+                            //очищаем только если есть изменения
+                            nodeService.removeAssociation(parent, existingAssoc.getTargetRef(), assocType);
+                        } else {
+                            //иначе оставляем имеющуюся ассоциацию
+                            create = false;
+                        }
 				    }
 			    }
 			    if (create) {
